requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pandas>=1.5.0
python-dotenv>=1.0.0
plotly>=5.15.0
wordcloud>=1.9.0
matplotlib>=3.7.0
numpy>=1.24.0
selenium>=4.15.0
webdriver-manager>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0